requirements.txt
aiohttp>=3.9.0
numpy>=1.26.0
pydantic>=2.5.0
python-dotenv>=1.0.0
sounddevice>=0.4.7
faster-whisper>=1.0.0
vosk>=0.3.45
websockets>=12.0
google-auth>=2.27.0
requests>=2.31.0
